+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Added this change log after project was already under way.  Early changes are no
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [2.1] - 2019-06-02
+### Changed
+- Switched to JUnit 5
+- Added new `fromUTF8` functions to `Strings` taking `Iterator` and `ByteBuffer`
+
 ## [2.0] - 2017-04-25
 ### Changed
 - Multiple classes modified for Java 8